Refactor accounts app ready method to import setup_new_trainer from signals
