+++ backend/botany/constants.py
@@ -1,0 +1,3 @@
+BOTANY_PERMISSIONS = [
+    'add_plant', 'change_plant', 'delete_plant', 'view_plant'
+]
