+++ backend/inventory/constants.py
@@ -1,0 +1,8 @@
+COLLECTION_PERMISSIONS = (
+    'add_image', 'change_image', 'choose_image',
+    'add_document', 'change_document', 'choose_document'
+)
+
+PAGE_PERMISSIONS = (
+    'add_page', 'publish_page'
+)
