+++ backend/inventory/utils.py
@@ -1,17 +1,12 @@
 from django.db import transaction
 
-from .models import InventoryIndexCollection, InventoryBoxPage
+from .models import InventoryBoxPage
 
 
 @transaction.atomic
-def setup_user_inventory(user):
-    user.add_to_group("Trainers")
-    user_collection = InventoryIndexCollection.get_for_user(user)
-    user_inventory = user_collection.get_user_page()
-    setup_inventory_boxes(user_inventory)
+def setup_inventory_boxes(user, num_boxes=5):
+    user_inventory = user.get_page()
 
-
-def setup_inventory_boxes(user_inventory, num_boxes=5):
     for i in range(1, num_boxes + 1):
         user_box_page = InventoryBoxPage(
             title=f"Box {i}",
